Update gift-membership-received.md
The docs for the gifter and receiver were opposite. The `gifter` is the person who purchased/gave the gifted membership. I also think it might be helpful to add the `user` and `userId` variables to be clear that those variables can be used for the user and ID of the receiver of the gifted membership.